[SO-083][FIX] Make cache telemetry async
