+++ lib/solid_observer/event_buffer_core.rb
@@ -1,0 +1,218 @@
+# frozen_string_literal: true
+
+require "concurrent/timer_task"
+
+module SolidObserver
+  module EventBufferCore
+    INITIAL_METRICS = {
+      flush_failures_count: 0,
+      drops_count: 0,
+      last_flush_at: nil,
+      last_flush_duration_ms: nil,
+      last_flush_error: nil
+    }.freeze
+
+    def initialize_event_buffer
+      @mutex = Mutex.new
+      @metrics_mutex = Mutex.new
+      @buffer = []
+      @metrics = INITIAL_METRICS.dup
+      @timer_task = nil
+    end
+    private :initialize_event_buffer
+
+    def push(event_data)
+      return unless (config = SolidObserver.config).persistence_mode?
+
+      drops_count, should_flush = sync_push_and_check(event_data, config)
+      record_drop(drops_count) if drops_count.positive?
+      ensure_timer_running
+      flush! if should_flush
+    end
+
+    def flush!
+      events_to_flush = drain_buffer
+      return if events_to_flush.empty?
+
+      flush_events(events_to_flush, monotonic_ms)
+    end
+
+    def flush
+      flush!
+    end
+
+    def size
+      @mutex.synchronize { @buffer.size }
+    end
+
+    def clear
+      @mutex.synchronize { @buffer.clear }
+    end
+
+    def metrics
+      current_size = @mutex.synchronize { @buffer.size }
+      snapshot = @metrics_mutex.synchronize { @metrics.dup }
+      {
+        size: current_size,
+        max_buffer_size: SolidObserver.config.max_buffer_size
+      }.merge(snapshot)
+    end
+
+    def shutdown
+      stop_timer
+      flush!
+    end
+
+    private
+
+    def drain_buffer
+      @mutex.synchronize do
+        events_to_flush = @buffer.dup
+        @buffer.clear
+        events_to_flush
+      end
+    end
+
+    def flush_events(events_to_flush, started_at_ms)
+      flush_service.call(events_to_flush)
+      record_flush_success(monotonic_ms - started_at_ms)
+    rescue => error
+      handle_flush_error(error, events_to_flush)
+    end
+
+    def handle_flush_error(error, events_to_flush)
+      requeue_failed_events(events_to_flush)
+      record_flush_failure(error)
+      log_flush_failure(error)
+      nil
+    end
+
+    def log_flush_failure(error)
+      Rails.logger&.error("[SolidObserver] #{log_label} flush failed: #{error.message}") if defined?(Rails)
+    end
+
+    def apply_overflow_policy(event_data, config)
+      if @buffer.size < config.max_buffer_size
+        @buffer << event_data
+        return 0
+      end
+
+      handle_overflow(event_data, config.buffer_overflow_strategy)
+    end
+
+    def handle_overflow(event_data, overflow_strategy)
+      case overflow_strategy
+      when :drop_old
+        @buffer.shift
+        @buffer << event_data
+      when :drop_new
+        nil
+      else
+        raise ArgumentError, "Unsupported buffer_overflow_strategy: #{overflow_strategy.inspect}"
+      end
+
+      1
+    end
+
+    def requeue_failed_events(events_to_flush)
+      dropped_count = sync_requeue_events(events_to_flush, SolidObserver.config.max_buffer_size)
+      record_drop(dropped_count) if dropped_count.positive?
+    end
+
+    def trim_events_for_capacity(events, max_buffer_size)
+      dropped_count = events.size - max_buffer_size
+      return [events, 0] if dropped_count <= 0
+
+      [events_to_keep(events, max_buffer_size), dropped_count]
+    end
+
+    def ensure_timer_running
+      timer_to_start, timer_to_stop = replace_timer_if_stopped
+      return unless timer_to_start
+
+      timer_to_stop&.shutdown
+      timer_to_start.execute
+    end
+
+    def replace_timer_if_stopped
+      @mutex.synchronize do
+        current_timer_task = @timer_task
+        return [nil, nil] if timer_running?(current_timer_task)
+
+        [build_timer_task, current_timer_task]
+      end
+    end
+
+    def stop_timer
+      timer_to_stop = @mutex.synchronize do
+        current_timer = @timer_task
+        @timer_task = nil
+        current_timer
+      end
+
+      timer_to_stop&.shutdown
+    end
+
+    def record_flush_success(duration_ms)
+      @metrics_mutex.synchronize do
+        @metrics.merge!(
+          last_flush_at: Time.current,
+          last_flush_duration_ms: duration_ms,
+          last_flush_error: nil
+        )
+      end
+    end
+
+    def record_flush_failure(error)
+      @metrics_mutex.synchronize do
+        @metrics[:flush_failures_count] += 1
+        @metrics[:last_flush_error] = error.message
+      end
+    end
+
+    def record_drop(count = 1)
+      @metrics_mutex.synchronize do
+        @metrics[:drops_count] += count
+      end
+    end
+
+    def sync_push_and_check(event_data, config)
+      @mutex.synchronize do
+        drops_count = apply_overflow_policy(event_data, config)
+        [drops_count, @buffer.size >= config.buffer_size]
+      end
+    end
+
+    def sync_requeue_events(events_to_flush, max_buffer_size)
+      @mutex.synchronize do
+        combined_events = events_to_flush + @buffer
+        kept_events, dropped_count = trim_events_for_capacity(combined_events, max_buffer_size)
+        @buffer.replace(kept_events)
+        dropped_count
+      end
+    end
+
+    def events_to_keep(events, max_buffer_size)
+      if SolidObserver.config.buffer_overflow_strategy == :drop_old
+        events.last(max_buffer_size)
+      else
+        events.first(max_buffer_size)
+      end
+    end
+
+    def timer_running?(timer_task)
+      timer_task && !timer_task.shuttingdown?
+    end
+
+    def build_timer_task
+      @timer_task = Concurrent::TimerTask.new(
+        execution_interval: SolidObserver.config.flush_interval,
+        run_now: false
+      ) { flush! }
+    end
+
+    def monotonic_ms
+      Process.clock_gettime(Process::CLOCK_MONOTONIC, :millisecond)
+    end
+  end
+end
